Spr 864 fs check utility rebased (#538)
* added new utility for extent validation

* added stats initialization

* added a read from stats table

* initial implementation of file system check utility

* fixed up some things

* create FSCheck class

* added secondary index validation

* forgot one file

* some cleanup

* code review fix

---------

+++ include/common/constants.hh
@@ -41,7 +41,7 @@ namespace springtail::constant {
     static constexpr uint32_t INDEX_DATA = std::numeric_limits<uint32_t>::max();
 
     /** An index ID that represents the primary index. */
-    static constexpr uint32_t INDEX_PRIMARY = 0;
+    static constexpr uint64_t INDEX_PRIMARY = 0;
 
     /** The name of the child pointer field in a BTree branch extent. */
     static const std::string BTREE_CHILD_FIELD = "__springtail_child";
